+++ app-backend/migrations/src/main/scala/migrations/Summary.scala
@@ -160,4 +160,5 @@ object MigrationSummary {
   M160
   M161
   M162
+  M163
 }